% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfectcircle.R
\name{circularity}
\alias{circularity}
\title{Measure circularity of a circle in an image}
\usage{
circularity(warp, seedPoints, progress, verbose = FALSE)
}
\arguments{
\item{warp}{The rectified image containing the circle}

\item{seedPoints}{A \code{data.frame} containing the (x, y) coordinates of \code{type=="foreground"} and \code{type=="background"} pixels.}

\item{progress}{A shiny progress bar.}
}
\value{
A list containing several elements
\describe{
   \item{outfile}{File name (with path) of the image containing both the extracted freehand circle and the best fitting circle}
   \item{outfile_intermediate}{File name (with path) containing the intermediate results}
   \item{area}{Area of the disc made up by the freehand circle}
   \item{ratio_area}{Ratio between the above \code{area} and the corresponding area of the perfect circle. 1 is optimial}
   \item{area_difference}{Summing up all points inside and outside the perfect circle weighted by distance (see blog post for details)}
   \item{score}{Corresponds to \code{(1-ratio_areadifference)*100}. The optimal value is 100\%.}
}
}
\description{
Function executes the steps described in \href{http://staff.math.su.se/hoehle/blog/2018/07/31/circle.html}{Judging Freehand Circle Drawing Competitions}.
Note: Opposite to the description in the blog post we assume that the image is already rectified.
}
\examples{
img <- imager::load.image(  system.file("extdata/", "localhost.jpg", package = "perfectcircle") )
seedPoints <- read.csv( system.file("extdata", "localhost.csv", package="perfectcircle"))
res <- perfectcircle::circularity(img, seedPoints, progress=NULL)
res

plot(imager::load.file(res$outfile))
}
\references{
Höhle, M. (2018), \href{http://staff.math.su.se/hoehle/blog/2018/07/31/circle.html}{Judging Freehand Circle Drawing Competitions}.
}
