% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfectcircle.R
\name{my_incProgress}
\alias{my_incProgress}
\title{Helper function to rectify an image - currently not used
since points have to be transmitted.}
\usage{
my_incProgress(progress, amount = 0.1, message = NULL, detail = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{img}{An image}

\item{p}{Control points
Helper - don't export}
}
\description{
Helper function to rectify an image - currently not used
since points have to be transmitted.
}
