% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfectcircle.R
\name{radius_given_center}
\alias{radius_given_center}
\title{Estimate best radius given center of the circle in an image.}
\usage{
radius_given_center(center, freehandCircleThinBorder, dist = NULL)
}
\arguments{
\item{center}{A vector of length two containing the (x,y) coordinates of the center}

\item{freehandCircleThinBorder}{An imager image mask with the extracted borderline of the freehand circle having a pixel value > 0}

\item{dist}{A distance matrix for each pixel with \code{freehandCircleThinBorder} > 0}
}
\value{
Returns the mean of the distance matrix, which is the radius of the circle fitting the points optimally.
}
\description{
Given a mask image containing the borderline of the circle and the center coordinates
one can compute for this configuration using the formula by Coope (1993).
}
\details{
The function implements Formula 3 in Coope (1993) while using a mask image to hold the points.
Given the center \eqn{c} we can determine the radius as \eqn{r(c)=\frac{1}{m} \sum_{j=1}^m ||c-a_j||_2}{r(c)=1/m \sum_{j=1}^m (c-a_j)^2}.
}
\references{
Coope, I.D. 1993. “Circle Fitting by Linear and Nonlinear Least Squares.” Journal of Optimization Theory and Applications 76 (2): 381–88. doi:10.1007/BF00939613.
}
