% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfectcircle.R
\name{target_tls}
\alias{target_tls}
\title{Target function of the total least squares criterion of Coope (1993)}
\usage{
target_tls(theta, freehandCircleThinBorder)
}
\arguments{
\item{theta}{A vector of length 2 containing log of the center coordinates.}
}
\value{
The total least squares \deqn{}{}
}
\description{
Target function of the total least squares criterion of Coope (1993)
}
\details{
Denoting by \eqn{c} the center of the circle and by \eqn{r>0} the radius we
want to find the circle for which the Euclidean distance to $m$ data points $a_j$, $j=1,\ldots,m$ using the following criterion:
\deqn{\min_{c\in \mathbb{R^2}, r>0} \sum_{j=1}^m F_j(c,r)^2}, where \deqn{F_j(c,r) = \left|r - ||c-a_j||_2\right|}.
This is equation (1) and (2) of Coope (1993).
}
\references{
Coope, I.D. 1993. “Circle Fitting by Linear and Nonlinear Least Squares.” Journal of Optimization Theory and Applications 76 (2): 381–88. doi:10.1007/BF00939613.
}
